Backend/requirements.txt
adrf==0.1.12
anyio==4.12.1
asgiref==3.9.1
async-property==0.2.2
cachetools==6.2.4
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
cloudinary==1.44.1
colorama==0.4.6
dj-database-url==3.1.0
Django==5.2.5
django-cloudinary-storage==0.3.0
django-cors-headers==4.9.0
django-environ==0.12.0
django-filter==25.2
django-js-asset==3.1.2
django-rest-passwordreset==1.5.0
django-taggit==6.1.0
djangorestframework==3.16.1
dnspython==2.8.0
email-validator==2.3.0
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.41.1
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.3
googleapis-common-protos==1.72.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httptools==0.7.1
httpx==0.28.1
idna==3.11
oauthlib==3.3.1
packaging==25.0
pillow==11.3.0
proto-plus==1.27.0
protobuf==6.33.2
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.6.2
uvicorn==0.40.0
watchfiles==1.1.1
websockets==16.0
whitenoise==6.11.0
